Added initial debruijn codes to avoid circular codes Delay replaced with code repeatition

+++ firmware/application/de_bruijn2.hpp
@@ -13,7 +13,7 @@ struct de_bruijn2 {
 
     bool abrt { false };
     void db(uint32_t t, uint32_t p, uint32_t n, uint32_t maxlen, uint16_t k, uint8_t* a, char* sequence, char* charset);
-    void db_callback(uint32_t t, uint32_t p, uint32_t n, uint32_t maxlen, uint16_t seqlen, uint16_t k, uint8_t* a, char* charset, std::function<void(char* seq_part, uint8_t seq_part_len)> callback);
+    void db_callback(uint32_t t, uint32_t p, uint32_t n, uint32_t maxlen, int32_t seqlen, uint16_t k, uint8_t* a, char* charset, std::function<void(char* seq_part, uint8_t seq_part_len)> callback);
 };
 
 #endif/*__DE_BRUIJN_H2__*/
